bbswitch_dev: manage bridge D3
nvidia.ko doesn't like its bridge gets put into D3. If the bridge is in
D3, it thinks the device drops off the bus.
So we need to disable bridge D3 before loading nvidia.ko to make it
happy.

We still want to use bridge D3 when ther bbswitch_dev is in charge,
hence enable it int probe function.